% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DevTimeSBW}
\alias{DevTimeSBW}
\title{DevTimeSBW}
\format{A list of data frames:
\describe{
\item{\code{IN.xx.F1}}{Development time of Inuvik sample F1 generation reared at xxºC}
\item{\code{NB.xx.F1}}{Development time of New Brunswick sample F1 generation reared at xxºC}
\item{\code{ON.xx.F1}}{Development time of Ontario sample F1 generation reared at xxºC}
\item{\code{QC.xx.F1}}{Development time of Quebec sample F1 generation reared at xxºC}
\item{\code{IPU.xx.F0}}{Development time of IPQL (Sault-ste-marie) sample F0 generation reared at xxºC}
\item{\code{IPU.xx.F1}}{Development time of IPQL (Sault-ste-marie) sample F1 generation reared at xxºC}
}

For further details, see Wardlaw et al. (in progress)}
\usage{
DevTimeSBW
}
\description{
Post-diapause development times of spruce budworm samples taken from different geographical locations
reared at different constant temperatures.
}
\keyword{datasets}
